Prototype Statistics & Profiling window
There are a few details missing from this mockups, see TODOs.

But it gets us a decent picture of what's possible or not. We can
improve from there.

This proposes a new "pinned stats" mechanism, which is what is
used to control what would be visible in the stats dock. The most
common stats are pinned by default.

The other innovation of this mockup is the redesign of the stats
window. The sidebar now reports the state of the page through icons,
so people can spot at a glance what needs their attention.

The audience of this redesigned stats dock and window are studios
and professional workflows, where constant monitoring of the stream
quality is required, and actions must be taken quickly.

Related: https://github.com/obsproject/design/issues/19